requirements.txt
pyserial==2.7
requests==2.7.0
